Rename key policy name file
