+++ lib/policies/key-auth/index.js
@@ -1,5 +1,5 @@
 module.exports = {
-  policy: require('./keyauth'),
+  policy: require('./key-auth'),
   schema: {
     $id: 'http://express-gateway.io/schemas/policies/key-auth.json',
     type: 'object',
