+++ lib/policies/key-auth/key-auth.js
